Initial checkin.
The goal of PersistentTranslationService is to provide a persistent
translation service to be used by Zope i18n framework (2.7+). It has to
decide how a (domain, msgid) is going to be translated.

This implementation indirects calls to different message catalogs, the
one implemented for now is Localizer. The code is organized to mimic the
interfaces used in Zope 3's translation service.